Remove accepts val from TransformingFetcher
